refactor: name to list item
