--- apps/journeys-admin/src/components/PageWrapper/NavigationDrawer/NavigationMenuItem/index.tsx
@@ -1,1 +1,0 @@
-export { NavigationMenuItem } from './NavigationMenuItem'
